feat(hotkeys): implement session hotkeys for renaming and stopping streams
