+++ apps/web/src/routes/session.$id.tsx
@@ -1,6 +1,7 @@
 import { createFileRoute } from '@tanstack/react-router';
 
 import { SessionPage } from '@/components/session/session-page';
+import { useActions } from '@/lib/actions';
 import { sessionQueryOptions, sessionMessagesInfiniteQueryOptions } from '@/lib/queries/chat';
 import {
   enabledProviderModelsQueryOptions,
@@ -8,6 +9,7 @@ import {
 } from '@/lib/queries/providers';
 import { queuedMessagesQueryOptions } from '@/lib/queries/queue';
 import { settingsQueryOptions } from '@/lib/queries/settings';
+import { useSessionHotkeys } from '@/lib/use-session-hotkeys';
 
 export const Route = createFileRoute('/session/$id')({
   loader: ({ context, params }) =>
@@ -24,5 +26,9 @@ export const Route = createFileRoute('/session/$id')({
 
 function SessionRouteComponent() {
   const { id } = Route.useParams();
+
+  const actions = useActions();
+  useSessionHotkeys(actions);
+
   return <SessionPage sessionId={id} />;
 }
